# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PrepSmart API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Recipe {
  id: number;
  userId: number;
  name: string;
  ingredients: string[];
  instructions: string;
  /** @nullable */
  prepTime: number | null;
  /** @nullable */
  calories: number | null;
  isFavorite: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateRecipeBody {
  name: string;
  ingredients: string[];
  instructions: string;
  /** @nullable */
  prepTime?: number | null;
  /** @nullable */
  calories?: number | null;
}

export interface UpdateRecipeBody {
  name?: string;
  ingredients?: string[];
  instructions?: string;
  /** @nullable */
  prepTime?: number | null;
  /** @nullable */
  calories?: number | null;
}

export interface MealPlan {
  id: number;
  userId: number;
  weekStart: string;
  createdAt: string;
}

export type MealWithRecipeMealType =
  (typeof MealWithRecipeMealType)[keyof typeof MealWithRecipeMealType];

export const MealWithRecipeMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface MealWithRecipe {
  id: number;
  mealPlanId: number;
  /** @nullable */
  recipeId: number | null;
  dayOfWeek: number;
  mealType: MealWithRecipeMealType;
  /** @nullable */
  customName: string | null;
  createdAt: string;
  recipe: Recipe | null;
}

export interface MealPlanWithMeals {
  id: number;
  userId: number;
  weekStart: string;
  createdAt: string;
  meals: MealWithRecipe[];
}

export interface CreateMealPlanBody {
  weekStart: string;
}

export type MealMealType = (typeof MealMealType)[keyof typeof MealMealType];

export const MealMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface Meal {
  id: number;
  mealPlanId: number;
  /** @nullable */
  recipeId: number | null;
  /** 0=Sunday, 1=Monday, ..., 6=Saturday */
  dayOfWeek: number;
  mealType: MealMealType;
  /** @nullable */
  customName: string | null;
  createdAt: string;
}

export type CreateMealBodyMealType =
  (typeof CreateMealBodyMealType)[keyof typeof CreateMealBodyMealType];

export const CreateMealBodyMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface CreateMealBody {
  /** @nullable */
  recipeId?: number | null;
  dayOfWeek: number;
  mealType: CreateMealBodyMealType;
  /** @nullable */
  customName?: string | null;
}

export type UpdateMealBodyMealType =
  (typeof UpdateMealBodyMealType)[keyof typeof UpdateMealBodyMealType];

export const UpdateMealBodyMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface UpdateMealBody {
  /** @nullable */
  recipeId?: number | null;
  dayOfWeek?: number;
  mealType?: UpdateMealBodyMealType;
  /** @nullable */
  customName?: string | null;
}

export interface GroceryList {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  mealPlanId: number | null;
  createdAt: string;
}

export interface GroceryItem {
  id: number;
  groceryListId: number;
  name: string;
  /** @nullable */
  quantity: string | null;
  isChecked: boolean;
}

export interface GroceryListWithItems {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  mealPlanId: number | null;
  createdAt: string;
  items: GroceryItem[];
}

export interface CreateGroceryItemBody {
  name: string;
  /** @nullable */
  quantity?: string | null;
}

export interface CreateGroceryListBody {
  name: string;
  /** @nullable */
  mealPlanId?: number | null;
  items?: CreateGroceryItemBody[];
}

export interface DashboardData {
  todayMeals: MealWithRecipe[];
  currentPlan: MealPlanWithMeals | null;
  recipeCount: number;
  favoriteCount: number;
  activeGroceryList: GroceryListWithItems | null;
  weeklyMealCount: number;
}

export type ListRecipesParams = {
  favoritesOnly?: boolean;
  search?: string;
};